requirements.txt
flask
flask-cors
spotipy
ytmusicapi
python-dotenv